Moved SerializeSpans and helper methods to spearate utils class.
